requirements.txt
fastapi
uvicorn
transformers
torch
librosa
pyannote.audio
streamlit
requests
pydub
python-dotenv
python-multipart